build ohlcv bars in base
